Implement RequestFactory instead of ServerRequestFactory
